refactor: validate if the uuid

+++ src/app/controllers/ContactController.js
@@ -99,6 +99,10 @@ class ContactController {
   async delete(request, response) {
     const { id } = request.params;
 
+    if(!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid user id' })
+    }
+
     await ContactsRepository.delete(id);
     // no content
     response.sendStatus(204);
